requirements.txt
asgiref==3.8.1
Django==5.1.6
django-environ==0.12.0
django-filter==24.3
django-json-api==1.2.1
djangorestframework==3.15.2
djangorestframework-jsonapi==7.1.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
